Refactor, delete unused code
